refactor: extract tournament business logic into service layer
Move bracket generation, match result handling, ELO updates, and filtered
film queries from the router into backend/services/tournament.py so the
router only handles HTTP concerns. Add tests for pure functions.
